svg/text/scripted/textlength-inconsistent.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:html="http://www.w3.org/1999/xhtml"><title>Inconsistent two &apos;textLength&apos; attributes</title><html:script src="/resources/testharness.js"/><html:script src="/resources/testharnessreport.js"/><html:link href="https://crbug.com/1399351" rel="help"/><g><text x="10" y="100" textLength="150"><tspan textLength="300">Scalable Vector</tspan></text></g><html:script>test(() =&gt; { const tspan = document.querySelector(&apos;tspan&apos;); assert_less_than(tspan.getStartPositionOfChar(0).x, tspan.getStartPositionOfChar(1).x); }, &apos;The first characters should be on the left of the second character.&apos;);</html:script></svg>
